create user and client. task not ready

+++ src/models/tables/Client.js
@@ -1,9 +1,9 @@
-import { sequelize } from "./connection";
-import { prop } from "./defaultProps";
+import { sequelize } from "../connection";
+import { prop } from "../defaultProps";
 
 const Client = sequelize.define("clients", {
   businessName: prop("STRING"),
-  cuit: prop("INTEGER"),
+  cuit: prop("BIGINT"),
   owner: prop("STRING"),
   franchises: prop("INTEGER"),
   active: prop("BOOLEAN", { defaultValue: true }),
